Preserve executable and encoding information in scripts with metadata
Fixes #241
