+++ HISTORY.rst
@@ -15,7 +15,7 @@ Release History
 - Invalid notebooks may cause a warning, but not a fatal error (#234)
 - Jupyter server extension leaves the contents manager unchanged if it is a sub-class of Jupytext's CM (#236)
 - Fixed format inference when metadata is present but not format information (#239)
-
+- Preserve executable and encoding information in scripts with metadata (#241)
 
 1.1.3 (2019-05-22)
 ++++++++++++++++++++++
